+++ blueledger-webapp/app/(protected)/expenses/edit/[id]/page.tsx
@@ -1,8 +1,8 @@
-import Expense from '@/mongoose/models/Expense';
+import Expense from '@/models/expense';
 import { notFound } from 'next/navigation';
 import React from 'react';
 import mongoose from 'mongoose';
-import ExpenseForm from '../../_components/expense-form';
+import ExpenseForm from '../../../../../components/expenses/expense-form';
 
 const EditExpensePage = async ({
   params,
