+++ blueledger-webapp/app/(protected)/expenses/new/page.tsx
@@ -1,4 +1,4 @@
-import ExpenseForm from '../_components/expense-form';
+import ExpenseForm from '../../../../components/expenses/expense-form';
 
 const NewExpensePage = () => {
   return <ExpenseForm />;
