+++ blueledger-webapp/app/(protected)/layout.tsx
@@ -1,4 +1,4 @@
-import NavBar from '@/components/custom/nav-bar/nav-bar';
+import NavBar from '@/components/nav-bar';
 
 const AuthLayout = ({ children }: { children: React.ReactNode }) => {
   return (
