+++ blueledger-webapp/components/nav-bar/index.ts
@@ -1,0 +1,3 @@
+import NavBar from './nav-bar';
+
+export default NavBar;
